+++ packages/asset-swapper/src/utils/market_operation_utils/types.ts
@@ -69,7 +69,6 @@ export enum ERC20BridgeSource {
     CurveV2 = 'Curve_V2',
     Lido = 'Lido',
     ShibaSwap = 'ShibaSwap',
-    Clipper = 'Clipper',
     // BSC only
     PancakeSwap = 'PancakeSwap',
     PancakeSwapV2 = 'PancakeSwap_V2',
@@ -455,6 +454,11 @@ export interface GetMarketOrdersOpts {
      * hopping to. E.g DAI->USDC via an adjacent token WETH
      */
     tokenAdjacencyGraph: TokenAdjacencyGraph;
+
+    /**
+     * Gas price to use for quote
+     */
+    gasPrice: BigNumber;
 }
 
 /**
@@ -534,10 +538,11 @@ export interface GenerateOptimizedOrdersOpts {
     bridgeSlippage?: number;
     maxFallbackSlippage?: number;
     excludedSources?: ERC20BridgeSource[];
-    feeSchedule?: FeeSchedule;
+    feeSchedule: FeeSchedule;
     exchangeProxyOverhead?: ExchangeProxyOverhead;
     allowFallback?: boolean;
     shouldBatchBridgeOrders?: boolean;
+    gasPrice: BigNumber;
 }
 
 export interface ComparisonPrice {