+++ contracts/treasury/CHANGELOG.md
@@ -5,6 +5,10 @@ Edit the package's CHANGELOG.json file only.
 
 CHANGELOG
 
+## v1.4.2 - _October 4, 2021_
+
+    * Dependencies updated
+
 ## v1.4.1 - _September 29, 2021_
 
     * Dependencies updated