+++ contracts/zero-ex/CHANGELOG.md
@@ -5,6 +5,10 @@ Edit the package's CHANGELOG.json file only.
 
 CHANGELOG
 
+## v0.29.1 - _October 4, 2021_
+
+    * Dependencies updated
+
 ## v0.29.0 - _September 29, 2021_
 
     * Export TransformERC20FeatureContract (#282)